+++ examples/babel-loader/test.sh
@@ -1,0 +1,3 @@
+diff ./happy/main.js ./vanilla/main.raw.js
+
+grep "success" ./happy/main.js