+++ examples/extract-text-webpack-plugin/src/index.js
@@ -1,8 +1,0 @@
-import React from 'react';
-import ReactDOM from 'react-dom';
-import App from './App.js';
-
-ReactDOM.render(
-  <App />,
-  document.getElementById('app')
-);